requirements.txt
asgiref==3.8.1
certifi==2026.2.25
charset-normalizer==3.4.7
django==4.2.29
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.11
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.4
sqlparse==0.5.5
typing-extensions==4.13.2
urllib3==2.2.3
gunicorn
openai